+++ CHANGES.md
@@ -1,3 +1,14 @@
+# Version 0.12.0 - January 14, 2025
+
+- Fixes https://github.com/open-quantum-safe/liboqs-python/issues/98. The API
+  that NIST has introduced in
+  [FIPS 204](https://csrc.nist.gov/pubs/fips/204/final)
+  for ML-DSA includes a context string of length >= 0. Added new API for
+  signing with a context string
+  - `Signature.sign_with_ctx_str(self, message, context)`
+  - `Signature.verify_with_ctx_str(self, message, signature, context,
+public_key)`
+
 # Version 0.10.0 - April 1, 2024
 
 - Replaced CHANGES by